filehandling ausgelagert
