+++ PWManager.CLI/Controllers/LoginController.cs
@@ -43,7 +43,7 @@ namespace PWManager.CLI.Controllers {
                 basepointer++;
             }
 
-            var lastUser = ReadDefaultFile();
+            var lastUser = ConfigFileHandler.ReadDefaultFile();
             if (String.IsNullOrWhiteSpace(username)) {
                 username = lastUser.Split('\n')[0];
             }
@@ -54,30 +54,11 @@ namespace PWManager.CLI.Controllers {
             var pass = Prompt.Password("Enter your password");
             _loginService.Login(username, pass, path);
 
-            WriteDefaultFile(username, path);
-
+            ConfigFileHandler.WriteDefaultFile(username, path);
             Console.WriteLine($"Welcome {username} :)");
+
+            _env.RunningSession = true;
             return ExitCondition.CONTINUE;
         }
-
-        private string ReadDefaultFile() {
-            var defaultFilePath = Path.GetDirectoryName(System.Reflection.Assembly.GetEntryAssembly().Location);
-
-            try {
-                return File.ReadAllText(Path.Combine(defaultFilePath, "last.txt"));
-            } catch (IOException e) {
-                throw new UserFeedbackException("The file could not be read");
-            }
-        }
-        private void WriteDefaultFile(string username, string path) {
-            var defaultFilePath = Path.GetDirectoryName(System.Reflection.Assembly.GetEntryAssembly().Location);
-
-            try {
-                File.WriteAllText(Path.Combine(defaultFilePath, "last.txt"), username + '\n' + path);
-            } catch (IOException e) {
-                throw new UserFeedbackException("The file could not be written");
-            }
-        }
-
     }
 }
